Add Makefile, libft.h, README, tests and migrate signatures to subject-compliant types
- Migrate ft_strlen/atoi/strdup/strlcpy/strlcat/strncmp/strnstr/substr to use size_t and const where the man page requires
- Move ft_memset from pendiente/ to root and clean ft_strdup (drop ft_size helper, reuse ft_strlen)
- Add libft.h with prototypes for all 18 implemented functions
- Add 42-style Makefile with -Wall -Wextra -Werror, targets all/clean/fclean/re
- Add tests/ with minunit, active tests for 18 functions and pending-test scaffolding for the remaining 16
- Add README.md per subject requirements (Description, Instructions, Resources, AI usage)
- Add .gitignore for .o files, .a archives and the test binary

+++ ft_substr.c
@@ -10,13 +10,13 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <stdlib.h>
+#include "libft.h"
 
-char	*ft_substr(char const *s, unsigned int start, unsigned int len)
+char	*ft_substr(char const *s, unsigned int start, size_t len)
 {
-	char			*result;
-	unsigned int	s_len;
-	unsigned int	i;
+	char	*result;
+	size_t	s_len;
+	size_t	i;
 
 	s_len = 0;
 	while (s[s_len])
